refactor(shell): own controlled curl execution
