+++ packages/workspace/src/ai.ts
@@ -205,289 +205,17 @@ async function runShellCommand(
   command: string,
   options: { broadSearchPaths: string[], commands: string[], cwd: string, maxOutputLength: number, timeout?: number },
 ): Promise<WorkspaceShellResult> {
-  const curlResult = await runSourceCurlCommand(input, command, options)
-  if (curlResult) return curlResult
-
+  const networkGrants = getWorkspaceSourceRequestExecution(input)
   const { createReadonlyWorkspaceFs, runWorkspaceInspectionCommand } = await import("@vite-hub/shell/workspace")
   return await runWorkspaceInspectionCommand(input, command, {
     broadSearchPaths: options.broadSearchPaths,
-    commands: options.commands,
+    commands: networkGrants ? [...options.commands, "curl"] : options.commands,
     cwd: options.cwd,
     fs: createReadonlyWorkspaceFs(input),
     maxOutputLength: options.maxOutputLength,
+    networkGrants,
     timeout: options.timeout,
   })
-}
-
-async function runSourceCurlCommand(
-  input: Workspace | WorkspaceAssets,
-  command: string,
-  options: { cwd: string, maxOutputLength: number, timeout?: number },
-): Promise<WorkspaceShellResult | undefined> {
-  if (!mentionsCurlCommand(command)) return undefined
-
-  const parsed = parseControlledCurlCommand(command)
-  if (!parsed.ok) return policyDeniedCurl(command, options.cwd, parsed.error)
-
-  const executor = getWorkspaceSourceRequestExecution(input)
-  if (!executor) {
-    return policyDeniedCurl(command, options.cwd, "No API-backed Source request descriptors are visible in this workspace.")
-  }
-
-  const started = Date.now()
-  try {
-    const result = await executor.executeSourceRequest({
-      body: parsed.body,
-      method: parsed.method,
-      url: parsed.url,
-    })
-    const stdout = typeof result.content === "string" ? result.content : new TextDecoder().decode(result.content)
-    const limited = limitShellOutput(stdout, options.maxOutputLength)
-    return {
-      command,
-      cwd: options.cwd,
-      durationMs: Date.now() - started,
-      event: "command_finished",
-      exitCode: 0,
-      maxOutputLength: options.maxOutputLength,
-      outputTruncated: limited.truncated,
-      stderr: "",
-      stdout: limited.output,
-    }
-  }
-  catch (error) {
-    const message = error instanceof Error ? error.message : String(error)
-    const httpFailure = message.includes("HTTP request failed")
-    return {
-      command,
-      cwd: options.cwd,
-      durationMs: Date.now() - started,
-      event: httpFailure ? "command_finished" : "policy_denied",
-      exitCode: httpFailure ? 22 : 126,
-      stderr: `${message}\n`,
-      stdout: "",
-    }
-  }
-}
-
-type ParsedControlledCurl =
-  | { body?: unknown, method: "GET" | "HEAD" | "POST", ok: true, url: string }
-  | { error: string, ok: false }
-
-function parseControlledCurlCommand(command: string): ParsedControlledCurl {
-  if (hasUnsupportedCurlShellSyntax(command)) {
-    return { error: "Controlled curl must be a single command without pipes, redirects, chaining, command substitution, or heredocs.", ok: false }
-  }
-
-  let words: string[]
-  try {
-    words = parseShellWords(command)
-  }
-  catch (error) {
-    return { error: error instanceof Error ? error.message : String(error), ok: false }
-  }
-
-  if (words[0] !== "curl") return { error: "Controlled curl command must start with curl.", ok: false }
-
-  let body: unknown
-  let method: "GET" | "HEAD" | "POST" | undefined
-  let url: string | undefined
-
-  try {
-    for (let index = 1; index < words.length; index += 1) {
-      const word = words[index]!
-      if (word === "-s" || word === "-S" || word === "-sS" || word === "--silent" || word === "--show-error" || word === "-L" || word === "--location") {
-        continue
-      }
-      if (word === "-X" || word === "--request") {
-        const value = words[++index]
-        if (!value) return { error: `${word} requires a method.`, ok: false }
-        method = parseCurlMethod(value)
-        if (!method) return { error: `Unsupported curl request method: ${value}.`, ok: false }
-        continue
-      }
-      if (word.startsWith("--request=")) {
-        method = parseCurlMethod(word.slice("--request=".length))
-        if (!method) return { error: `Unsupported curl request method: ${word.slice("--request=".length)}.`, ok: false }
-        continue
-      }
-      if (word === "--url") {
-        const value = words[++index]
-        if (!value) return { error: "--url requires a URL.", ok: false }
-        url = setCurlUrl(url, value)
-        continue
-      }
-      if (word.startsWith("--url=")) {
-        url = setCurlUrl(url, word.slice("--url=".length))
-        continue
-      }
-      if (word === "--json") {
-        const value = words[++index]
-        if (!value) return { error: "--json requires a JSON body.", ok: false }
-        const parsed = parseJsonCurlBody(value)
-        if (!parsed.ok) return parsed
-        body = parsed.body
-        method ??= "POST"
-        continue
-      }
-      if (word.startsWith("--json=")) {
-        const parsed = parseJsonCurlBody(word.slice("--json=".length))
-        if (!parsed.ok) return parsed
-        body = parsed.body
-        method ??= "POST"
-        continue
-      }
-      if (word === "-d" || word === "--data" || word === "--data-raw" || word === "--data-binary" || word.startsWith("-d") || word.startsWith("--data=") || word.startsWith("--data-raw=") || word.startsWith("--data-binary=")) {
-        return { error: "Controlled curl v1 does not allow -d/--data flags. Use --json with a declared bodySchema.", ok: false }
-      }
-      if (word === "-H" || word === "--header" || word === "-b" || word === "--cookie" || word.startsWith("--header=") || word.startsWith("--cookie=")) {
-        return { error: "Controlled curl injects Source credentials itself; do not pass headers or cookies in the command.", ok: false }
-      }
-      if (word.startsWith("-")) return { error: `Unsupported curl flag: ${word}.`, ok: false }
-      url = setCurlUrl(url, word)
-    }
-  }
-  catch (error) {
-    return { error: error instanceof Error ? error.message : String(error), ok: false }
-  }
-
-  if (!url) return { error: "Controlled curl requires a URL.", ok: false }
-  try {
-    url = new URL(url).toString()
-  }
-  catch {
-    return { error: `Controlled curl URL is invalid: ${url}.`, ok: false }
-  }
-
-  return {
-    body,
-    method: method ?? "GET",
-    ok: true,
-    url,
-  }
-}
-
-function mentionsCurlCommand(command: string): boolean {
-  return parseShellWordsLenient(command)[0] === "curl"
-}
-
-function hasUnsupportedCurlShellSyntax(command: string): boolean {
-  let quote: "'" | "\"" | undefined
-  let escaped = false
-  for (let index = 0; index < command.length; index += 1) {
-    const char = command[index]!
-    const next = command[index + 1]
-    if (escaped) {
-      escaped = false
-      continue
-    }
-    if (char === "\\") {
-      escaped = true
-      continue
-    }
-    if (quote) {
-      if (char === quote) quote = undefined
-      continue
-    }
-    if (char === "'" || char === "\"") {
-      quote = char
-      continue
-    }
-    if (char === "|" || char === ";" || char === "<" || char === ">" || char === "`" || char === "$" && next === "(") return true
-    if ((char === "&" || char === "|") && next === char) return true
-  }
-  return /<<-?/.test(command)
-}
-
-function parseShellWordsLenient(command: string): string[] {
-  try {
-    return parseShellWords(command)
-  }
-  catch {
-    return command.trim().split(/\s+/).filter(Boolean)
-  }
-}
-
-function parseShellWords(command: string): string[] {
-  const words: string[] = []
-  let current = ""
-  let quote: "'" | "\"" | undefined
-  let escaped = false
-
-  for (const char of command) {
-    if (escaped) {
-      current += char
-      escaped = false
-      continue
-    }
-    if (char === "\\") {
-      escaped = true
-      continue
-    }
-    if (quote) {
-      if (char === quote) quote = undefined
-      else current += char
-      continue
-    }
-    if (char === "'" || char === "\"") {
-      quote = char
-      continue
-    }
-    if (/\s/.test(char)) {
-      if (current) {
-        words.push(current)
-        current = ""
-      }
-      continue
-    }
-    current += char
-  }
-
-  if (escaped) current += "\\"
-  if (quote) throw new Error("unterminated quote")
-  if (current) words.push(current)
-  return words
-}
-
-function parseCurlMethod(value: string): "GET" | "HEAD" | "POST" | undefined {
-  const method = value.toUpperCase()
-  return method === "GET" || method === "HEAD" || method === "POST" ? method : undefined
-}
-
-function parseJsonCurlBody(value: string): { body: unknown, ok: true } | { error: string, ok: false } {
-  try {
-    return { body: JSON.parse(value), ok: true }
-  }
-  catch {
-    return { error: "--json body must be valid JSON.", ok: false }
-  }
-}
-
-function setCurlUrl(current: string | undefined, next: string): string {
-  if (current) throw new Error("Controlled curl accepts exactly one URL.")
-  return next
-}
-
-function policyDeniedCurl(command: string, cwd: string, message: string): WorkspaceShellResult {
-  return {
-    command,
-    cwd,
-    event: "policy_denied",
-    exitCode: 126,
-    stderr: `[vitehub] ${message}\n`,
-    stdout: [
-      "[vitehub] Controlled curl request was not run.",
-      "Inspect `.vitehub/sources/<sourceKey>.json`, then retry with a single curl command that matches a visible Source request descriptor.",
-    ].join("\n") + "\n",
-  }
-}
-
-function limitShellOutput(output: string, maxOutputLength: number): { output: string, truncated?: boolean } {
-  if (output.length <= maxOutputLength) return { output }
-  return {
-    output: `${output.slice(0, maxOutputLength)}\n[output truncated to ${maxOutputLength} characters]\n`,
-    truncated: true,
-  }
 }
 
 function sizeOf(content: string | Uint8Array) {
